requirements.txt
# 如果不用docker image
# torch
# torchvision

# main.py 原本有用到的
tqdm
catboost
mediapipe
scenedetect

# llm及調用需要
fastapi
uvicorn[standard]
python-multipart
opencv-python
pillow
ultralytics
ollama
supervision
python-dotenv

